move model output formatting functions to separate module
